move methods from UserService to AuthService
